rename registers namespace
