+++ src/Lego/Foundation/Button.php
@@ -1,7 +1,7 @@
 <?php namespace Lego\Foundation;
 
 use Collective\Html\HtmlFacade;
-use Lego\Register\Data\HighPriorityResponse;
+use Lego\Register\HighPriorityResponse;
 
 /**
  * Html button or link.
@@ -101,7 +101,6 @@ class Button
 
     public function action(\Closure $action)
     {
-        /** @var HighPriorityResponse $resp */
         $resp = lego_register(HighPriorityResponse::class, $action, md5('button ' . $this->text));
         return $this->url($resp->url());
     }
